pyproject.toml
[project]
name = "ciphered-pattern-api"
version = "0.1.0"
description = "API for patterns within simple substitution ciphers"
requires-python = ">=3.8"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.12",
    "fastapi-cache2>=0.2.2",
    "pydantic>=2.10.6",
    "tqdm>=4.67.1",
    "uvicorn>=0.33.0",
]

